[ADD] sale_stock_sn_single
For S/N tracked products, ensures that pickings moves created from a sale stock
are created with 1.0 quantity each one.

This is specially useful for e.g. manufacture stock rules, that could create
MOs with more than 1 to produce quantity if not, that is actually incorrect for
s/n products.